refactor(api): move stage-collection to new arbo
