+++ api/lib/routes.js
@@ -38,7 +38,7 @@ import * as passwords from './application/passwords/index.js';
 import * as poleEmploi from './application/pole-emploi/index.js';
 import * as progressions from './application/progressions/index.js';
 import * as saml from './application/saml/index.js';
-import * as stageCollection from './application/stage-collections/index.js';
+import * as stageCollection from '../src/evaluation/application/stage-collections/index.js';
 import * as scoringSimulator from './application/scoring-simulator/index.js';
 import * as scorecards from './application/scorecards/index.js';
 import * as scoOrganizationLearners from './application/sco-organization-learners/index.js';
