+++ api/scripts/prod/target-profile-migrations/revert-stages-to-threshold.js
@@ -13,7 +13,7 @@ import { learningContentCache as cache } from '../../../lib/infrastructure/cache
 import { disconnect } from '../../../db/knex-database-connection.js';
 
 import * as targetProfileForAdminRepository from '../../../lib/infrastructure/repositories/target-profile-for-admin-repository.js';
-import * as stageCollectionRepository from '../../../lib/infrastructure/repositories/target-profile-management/stage-collection-repository.js';
+import * as stageCollectionRepository from '../../../src/evaluation/infrastructure/repositories/stage-collection-repository.js';
 
 set_fs(fs);
