+++ api/src/evaluation/application/stage-collections/index.js
@@ -1,7 +1,7 @@
 import Joi from 'joi';
-import { securityPreHandlers } from '../security-pre-handlers.js';
+import { securityPreHandlers } from '../../../../lib/application/security-pre-handlers.js';
 import { stageCollectionController } from './stage-collection-controller.js';
-import { identifiersType } from '../../domain/types/identifiers-type.js';
+import { identifiersType } from '../../../../lib/domain/types/identifiers-type.js';
 
 const register = async function (server) {
   server.route([
