+++ api/src/evaluation/domain/usecases/index.js
@@ -7,6 +7,7 @@ import * as answerRepository from '../../infrastructure/repositories/answer-repo
 import * as assessmentRepository from '../../../shared/infrastructure/repositories/assessment-repository.js';
 import * as competenceRepository from '../../../shared/infrastructure/repositories/competence-repository.js';
 import * as competenceEvaluationRepository from '../../infrastructure/repositories/competence-evaluation-repository.js';
+import * as stageCollectionForTargetProfileRepository from '../../infrastructure/repositories/stage-collection-repository.js';
 import * as feedbackRepository from '../../infrastructure/repositories/feedback-repository.js';
 import * as stageRepository from '../../../../lib/infrastructure/repositories/stage-repository.js';
 import * as targetProfileForAdminRepository from '../../../../lib/infrastructure/repositories/target-profile-for-admin-repository.js';
@@ -23,6 +24,7 @@ const dependencies = {
   assessmentRepository,
   competenceEvaluationRepository,
   competenceRepository,
+  stageCollectionForTargetProfileRepository,
   feedbackRepository,
   stageRepository,
   targetProfileForAdminRepository,
