+++ api/src/evaluation/infrastructure/repositories/stage-collection-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../../db/knex-database-connection.js';
-import { StageCollection } from '../../../domain/models/target-profile-management/StageCollection.js';
+import { StageCollection } from '../../../../lib/domain/models/target-profile-management/StageCollection.js';
 
 const getByTargetProfileId = async function (targetProfileId) {
   const stages = await knex('stages').where({ targetProfileId }).orderBy('id', 'asc');
