+++ api/tests/evaluation/acceptance/application/stage-collections/stage-collection-controller_test.js
@@ -5,8 +5,8 @@ import {
   generateValidRequestAuthorizationHeader,
   learningContentBuilder,
   mockLearningContent,
-} from '../../../test-helper.js';
-import { createServer } from '../../../../server.js';
+} from '../../../../test-helper.js';
+import { createServer } from '../../../../../server.js';
 
 describe('Acceptance | Controller | stage-collection', function () {
   let server;
