+++ api/tests/integration/infrastructure/repositories/target-profile-management/stage-collection-repository_test.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { expect, databaseBuilder, domainBuilder, knex } from '../../../../test-helper.js';
-import * as stageCollectionRepository from '../../../../../lib/infrastructure/repositories/target-profile-management/stage-collection-repository.js';
+import * as stageCollectionRepository from '../../../../../src/evaluation/infrastructure/repositories/stage-collection-repository.js';
 import { StageCollectionUpdate } from '../../../../../lib/domain/models/target-profile-management/StageCollectionUpdate.js';
 
 describe('Integration | Infrastructure | Repository | target-profile-management | stage-collection-repository', function () {
